Pass error handler to response validator (#22)
* Fixed null references and internal shcemas
* Changed the body validation to match OpenAPI 2.0
spec and changed how the tests run

https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#schemaObject

Body - The payload that's appended to the HTTP request. Since there can only be one payload, there can only be one body parameter. The name of the body parameter has no effect on the parameter itself and is used for documentation purposes only. Since Form parameters are also in the payload, body and form parameters cannot exist together for the same operation

* Added responses  handling

* Fixed merge

* Allowed users to call next with error to propagate the error to the next error handler

* Request validator was not passing next to handlers